reorganized project, updated tests, works with latest Jive API
