+++ lib/jive/gettable_binary_url.rb
@@ -1,0 +1,14 @@
+require 'net/http'
+
+module Jive
+  module GettableBinaryURL
+    def get_binary
+      binary_uri = URI @binary_url
+      http = Net::HTTP.new(binary_uri.host, binary_uri.port)
+      req = Net::HTTP::Get.new binary_uri.request_uri
+      req.basic_auth @api_instance.auth[:username], @api_instance.auth[:password]
+      response = http.request req
+      response.body
+    end
+  end
+end
