+++ lib/jive/version.rb
@@ -1,0 +1,3 @@
+module Jive
+  VERSION = '0.0.7.pre'
+end
